+++ spec/rails_app/app/controllers/pages_controller.rb
@@ -1,23 +1,21 @@
 class PagesController < ApplicationController
+  respond_to :html, :js, :only => :about
+
   def readme
     @readme = IO.read(File.expand_path("../../../../../README.md",__FILE__))
     respond_to do |format|
       format.html
-      ajax_pagination format, :pagination => :menu, :partial => {:file => "pages/readme"}
+      ajax_pagination format, :pagination => :menu, :render => {:file => "pages/readme"}
     end
   end
 
   def about
-    respond_to do |format|
-      format.html
-      ajax_pagination format, :pagination => :menu, :partial => {:file => "pages/about"}
-    end
   end
 
   def welcome
     respond_to do |format|
       format.html
-      ajax_pagination format, :pagination => :menu, :partial => {:file => "pages/welcome"}
+      ajax_pagination format, :pagination => :menu
     end
   end
 end